this place too

+++ pycbc/events/stat.py
@@ -506,7 +506,7 @@ class PhaseTDStatistic(QuadratureSumStatistic):
 
             # renormalise to PDF
             self.weights[ifo] = \
-                weights[ifo] / (weights[ifo].sum() * bin_volume)
+                (weights[ifo] / (weights[ifo].sum() * bin_volume)).astype(numpy.float32)
 
             if param[ifo].dtype == numpy.int8:
                 # Older style, incorrectly sorted histogram file
